add AutoPredictor from_pretrained
