+++ paddlenlp/inference/__init__.py
@@ -12,4 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from .config import GenerationConfig
 from .predictor import AutoPredictor
